# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lottery/macau:
    get:
      operationId: getLotteryData
      tags: [lottery]
      summary: Get Toto Macau lottery data for a year
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            enum: [2025, 2026]
      responses:
        "200":
          description: Lottery data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LotteryYearData"
...
  /lottery/macau/stats:
    get:
      operationId: getLotteryStats
      tags: [lottery]
      summary: Basic stats for a year
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            enum: [2025, 2026]
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LotteryStats"
...
  /analytics/advanced:
    get:
      operationId: getAdvancedAnalytics
      tags: [analytics]
      summary: Full advanced analytics
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
            enum: [2025, 2026]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 500
      responses:
        "200":
          description: Advanced analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvancedAnalytics"
  /analytics/heatmap:
    get:
      operationId: getHeatmap
      tags: [analytics]
      summary: Position-digit heatmap data
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
            enum: [2025, 2026]
      responses:
        "200":
          description: Heatmap data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeatmapData"
  /analytics/markov:
    get:
      operationId: getMarkovChain
      tags: [analytics]
      summary: Markov chain transition matrix
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
            enum: [2025, 2026]
      responses:
        "200":
          description: Markov chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkovData"
...
components:
  schemas:
...
    DrawResult:
      type: object
      properties:
        date:
          type: string
        day:
          type: string
        s0001:
          type: ["string", "null"]
        s1300:
          type: ["string", "null"]
        s1600:
          type: ["string", "null"]
        s1900:
          type: ["string", "null"]
        s2200:
          type: ["string", "null"]
        s2300:
          type: ["string", "null"]
      required: [date, day, s0001, s1300, s1600, s1900, s2200, s2300]
    MonthData:
      type: object
      properties:
        month:
          type: string
        monthNumber:
          type: integer
        year:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/DrawResult"
      required: [month, monthNumber, year, results]
    LotteryYearData:
      type: object
      properties:
        year:
          type: integer
        months:
          type: array
          items:
            $ref: "#/components/schemas/MonthData"
        totalDraws:
          type: integer
        lastUpdated:
          type: string
      required: [year, months, totalDraws, lastUpdated]
...
    NumberFrequency:
      type: object
      properties:
        number:
          type: string
        count:
          type: integer
      required: [number, count]
    LotteryStats:
      type: object
      properties:
        year:
          type: integer
        totalDraws:
          type: integer
        mostFrequent:
          type: array
          items:
            $ref: "#/components/schemas/NumberFrequency"
        leastFrequent:
          type: array
          items:
            $ref: "#/components/schemas/NumberFrequency"
        digitFrequency:
          type: object
          additionalProperties:
            type: integer
      required: [year, totalDraws, mostFrequent, leastFrequent, digitFrequency]
...
    DigitFreqEntry:
      type: object
      properties:
        digit:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required: [digit, count, percentage]
    PairFreqEntry:
      type: object
      properties:
        pair:
          type: string
        count:
          type: integer
      required: [pair, count]
    GapEntry:
      type: object
      properties:
        number:
          type: string
        lastSeen:
          type: string
        gapDraws:
          type: integer
      required: [number, lastSeen, gapDraws]
    AdvancedAnalytics:
      type: object
      properties:
        totalDraws:
          type: integer
        digitFrequency:
          type: array
          items:
            $ref: "#/components/schemas/DigitFreqEntry"
        pairFrequency:
          type: array
          items:
            $ref: "#/components/schemas/PairFreqEntry"
        hotNumbers:
          type: array
          items:
            $ref: "#/components/schemas/NumberFrequency"
        coldNumbers:
          type: array
          items:
            $ref: "#/components/schemas/NumberFrequency"
        oddEven:
          type: object
          properties:
            odd:
              type: integer
            even:
              type: integer
            oddPct:
              type: number
            evenPct:
              type: number
          required: [odd, even, oddPct, evenPct]
        bigSmall:
          type: object
          properties:
            big:
              type: integer
            small:
              type: integer
            bigPct:
              type: number
            smallPct:
              type: number
          required: [big, small, bigPct, smallPct]
        missingNumbers:
          type: array
          items:
            type: string
        totalMissingCount:
          type: integer
        gapDistribution:
          type: array
          items:
            $ref: "#/components/schemas/GapEntry"
        repeatNumbers:
          type: array
          items:
            $ref: "#/components/schemas/NumberFrequency"
        trendNumbers:
          type: array
          items:
            $ref: "#/components/schemas/NumberFrequency"
      required:
        - totalDraws
        - digitFrequency
        - pairFrequency
        - hotNumbers
        - coldNumbers
        - oddEven
        - bigSmall
        - missingNumbers
        - totalMissingCount
        - gapDistribution
        - repeatNumbers
        - trendNumbers
    HeatmapData:
      type: object
      properties:
        positions:
          type: array
          items:
            type: object
            properties:
              position:
                type: integer
              label:
                type: string
              digits:
                type: array
                items:
                  type: object
                  properties:
                    digit:
                      type: string
                    count:
                      type: integer
                    pct:
                      type: number
                  required: [digit, count, pct]
            required: [position, label, digits]
        totalNumbers:
          type: integer
      required: [positions, totalNumbers]
    MarkovData:
      type: object
      properties:
        matrix:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: number
        topTransitions:
          type: array
          items:
            type: object
            properties:
              from:
                type: string
              to:
                type: string
              probability:
                type: number
            required: [from, to, probability]
        totalTransitions:
          type: integer
      required: [matrix, topTransitions, totalTransitions]
...